+++ coin/src/components/units/main/widget/CryptoWidget/CryptoWidget.tsx
