+++ coin/src/components/units/main/widget/DataWidget/DataWidgetContent.tsx
